+++ libs/components/src/lib/accordion-item/accordion-item.ts
@@ -1,7 +1,19 @@
 import { attr } from '@microsoft/fast-element';
 import { applyMixins, AccordionItem as FASTAccordionItem } from '@microsoft/fast-foundation';
 import { AffixIconWithTrailing } from '../../shared/patterns/affix';
+import type { Size } from '../enums.js';
 
+
+/**
+ * Types of accordion size.
+ *
+ * @public
+ */
+export type AccordionItemSize = Extract<Size, Size.Condensed | Size.Normal>;
+
+/**
+ * @slot - Default slot.
+ */
 export class AccordionItem extends FASTAccordionItem {
 	/**
 	 *
@@ -27,7 +39,20 @@ export class AccordionItem extends FASTAccordionItem {
 	 * HTML Attribute: meta
 	 */
 	@attr meta?: string;
+
+	/**
+	 * The size the accordion-item should have.
+	 *
+	 * @public
+	 * @remarks
+	 * HTML Attribute: size
+	 */
+	@attr size?: AccordionItemSize;
 }
+
+
+
+
 
 export interface AccordionItem extends AffixIconWithTrailing { }
 applyMixins(AccordionItem, AffixIconWithTrailing);