+++ libs/components/src/lib/action-group/action-group.ts
@@ -28,6 +28,7 @@ export type ActionGroupShape = Extract<Shape, Shape.Rounded | Shape.Pill>;
  * Base class for action-group
  *
  * @public
+ * @slot - Default slot.
  */
 export class ActionGroup extends FoundationElement {
 	/**
@@ -59,6 +60,8 @@ export class ActionGroup extends FoundationElement {
 		mode: 'boolean',
 	}) tight = false;
 
+	@attr() role: string | undefined;
+	@attr({attribute: 'aria-label'}) override ariaLabel: string | null = null;
 }