+++ libs/components/src/lib/button/button.ts
@@ -37,7 +37,7 @@ export type ButtonShape = Extract<Shape, Shape.Rounded | Shape.Pill>;
  *
  * @public
  */
-export type ButtonSize = Extract<Size, Size.Condensed | Size.Normal | Size.Expanded>;
+export type ButtonSize = Extract<Size, Size.SuperCondensed | Size.Condensed | Size.Normal | Size.Expanded>;
 
 /**
  * Base class for button
@@ -92,6 +92,18 @@ export class Button extends FoundationButton {
 		mode: 'boolean',
 		attribute: 'stacked',
 	}) stacked = false;
+
+	/**
+	 * Displays the button in pending state.
+	 *
+	 * @public
+	 * @remarks
+	 * HTML Attribute: pending
+	 */
+	@attr({
+		mode: 'boolean',
+		attribute: 'pending',
+	}) pending = false;
 
 	/**
 	 * Indicates the button's label.