+++ libs/components/src/lib/dialog/dialog.ts
@@ -23,6 +23,10 @@ export type IconPlacement = 'top' | 'side';
  * Base class for dialog
  *
  * @public
+ * @slot graphic - Use the graphic slot in order to replace the icon.
+ * @slot body - Use the body slot in order to add custom HTML to the dialog.
+ * @slot footer - Use the footer slot in order to add action buttons to the bottom of the dialog.
+ * @slot main - Assign nodes to the main slot to fully override a dialog’s predefined flow and style with your own.
  */
 export class Dialog extends FoundationElement {
 	/**