+++ libs/components/src/lib/tag-group/index.ts
@@ -1,0 +1,3 @@
+import { registerTagGroup } from './definition';
+
+registerTagGroup();