+++ libs/components/src/lib/tag-group/ui.test.ts
@@ -1,0 +1,33 @@
+import * as path from 'path';
+import { expect, test } from '@playwright/test';
+import type { Page } from '@playwright/test';
+import {
+	extractHTMLBlocksFromReadme,
+	loadComponents,
+	loadTemplate,
+} from '../../visual-tests/visual-tests-utils.js';
+
+const components = ['tag-group', 'tag'];
+
+test('should show the component', async ({ page }: { page: Page }) => {
+	const template = '<style>#wrapper{height: 250px; width: 1600px; display: flex; flex-wrap: wrap;}</style>' + extractHTMLBlocksFromReadme(
+		path.join(new URL('.', import.meta.url).pathname, 'README.md'))
+		.reduce((htmlString: string, block: string) => `${htmlString} <div style="margin: 5px;">${block}</div>`, '');
+
+	await loadComponents({
+		page,
+		components,
+	});
+	await loadTemplate({
+		page,
+		template,
+	});
+
+	const testWrapper = await page.$('#wrapper');
+
+	await page.waitForLoadState('networkidle');
+
+	expect(await testWrapper?.screenshot()).toMatchSnapshot(
+		'./snapshots/tag-group.png'
+	);
+});